Add connection schema RFID to RC522. Add autoconnect to wifi.

+++ src/ShoppingListArduino/ShoppingListArduino/API/CommonController.cs
@@ -117,6 +117,8 @@ namespace ShoppingListArduino.API
         [Route("add-unassigned-rfid")]
         public JObject AddUnassignedRfid(string rfid, string userId)
         {
+            if(_context.UserProductRfids.Any(x => x.Rfid == rfid))
+                return JObject.FromObject(new { success = false, message = "Продукт с такой меткой уже есть в базе." });
             try
             {
                 var result = _cache.Set(userId, rfid, TimeSpan.FromMinutes(20.0));
@@ -135,6 +137,11 @@ namespace ShoppingListArduino.API
         public JObject BindUnassignedRfidToUserProduct(int userProductId)
         {
             var userProduct = _context.UserProducts.Find(userProductId);
+
+            if (_context.UserProductRfids.Count(x=>x.UserProductId==userProductId) == userProduct.Quantity)
+            {
+                return JObject.FromObject(new { success = false, message = "Все товары данного типа уже привязаны к RFID метке" });
+            }
             string unassignedRfid;
             if (_cache.TryGetValue(userProduct.UserId, out unassignedRfid))
             {
@@ -168,7 +175,9 @@ namespace ShoppingListArduino.API
         [Route("user-product-to-bin")]
         public JObject UserProductToBin(string userId, int productId, int quantity)
         {
-            var userProduct = _context.UserProducts.FirstOrDefault(p => p.UserId == userId && p.ProductId == productId);
+            var userProduct = _context.UserProducts.
+                Include(x => x.UserProductRfids)
+                .FirstOrDefault(p => p.UserId == userId && p.ProductId == productId);
             if (userProduct == null)
             {
                 return JObject.FromObject(new { success = false, message = "У Вас дома уже не числится такой продукт!" });
@@ -176,7 +185,10 @@ namespace ShoppingListArduino.API
                 userProduct.Quantity -= quantity;
             if (userProduct.Quantity <= 0)
             {
-                _context.UserProductRfids.RemoveRange(userProduct.UserProductRfids);
+                if (userProduct.UserProductRfids.Count > 0)
+                {
+                    _context.UserProductRfids.RemoveRange(userProduct.UserProductRfids);
+                }
                 _context.UserProducts.Remove(userProduct);
             }
             try
